Boundaries + PVP
If you fall off the map you will now be respawned back onto the map with a death to the death counter.

Players can now shoot each other to occur deaths but will need to improve on the multiplayer part to ensure it is all working as intended between what you see on both screens.