Fixing Main Menu
Edited the Main Menu code to change to the right scene.

Modified the Menu Scene that is located in Assets > Scenes so that is has a Start Menu Manager that has the Main Menu script applied. This means you just drag that object into the game object area in the inspector to get the option functionality.